feat: implement secondary sidebar and empty page component; refactor routing and remove direct chats list
