+++ client/src/app/router/Routes.tsx
@@ -16,7 +16,7 @@ import JoinChatRoomPage from "../../features/chatRooms/join/JoinChatRoomPage";
 import ProfilePage from "../../features/profile/ProfilePage";
 import FriendsList from "../../features/friends/FriendsList";
 import AuthLayout from "../layout/AuthLayout";
-import DirectChatsList from "../../features/directChats/DirectChatList";
+import EmptyPage from "../layout/EmptyPage";
 import DirectChatDetails from "../../features/directChats/DirectChatDetails";
 
 export const router = createBrowserRouter([
@@ -27,14 +27,13 @@ export const router = createBrowserRouter([
       {
         element: <RequireAuth />,
         children: [
-          { path: "", element: <Navigate to="/direct-chats" replace /> },
+          { path: "", element: <EmptyPage /> },
           { path: "chat-rooms/:id", element: <ChatRoomDetails /> },
           { path: "create-chat-room", element: <ChatRoomForm key="create" /> },
           { path: "chat-rooms/:id/:token/join", element: <JoinChatRoomPage /> },
           { path: "manage/:id", element: <ChatRoomForm /> },
           { path: "profiles/:id", element: <ProfilePage /> },
           { path: "friends", element: <FriendsList /> },
-          { path: "direct-chats", element: <DirectChatsList /> },
           { path: "direct-chats/:id", element: <DirectChatDetails /> },
         ],
       },
